修改数据源

+++ Module02/page_energy/wrapped/func00_function.py
@@ -306,7 +306,6 @@ def data_deal_num_2(data_df,refer_df,flag):
         # stats_result = pd.concat((stats_result, tmp_df), axis=0)
         stats_result.insert(loc=0, column='时间', value=stats_result.index)
         stats_result.reset_index(drop=True, inplace=True)
-        
         # stats_result.at[0,'区域均值1'] = '开始日期'
         # stats_result.at[0,'区域均值2'] = '结束日期'
         # stats_result.at[0,'区域最大值1'] = '开始日期'
@@ -318,7 +317,7 @@ def data_deal_num_2(data_df,refer_df,flag):
         # stats_result.at[0,'区域距平百分率1'] = '开始日期'
         # stats_result.at[0,'区域距平百分率2'] = '结束日期'
         stats_result = stats_result.drop(0)
-
+        stats_result['时间']=stats_result['时间'].astype(int)
     return stats_result
 
 def calculate_average_hd(pre_data,ele):
